Use AnimatedNavHost in terminate insurance feature
Make use of our navigation:navigation-compose-typed module which exposes
 the accompanist backed navigation functions.
Also use DS motion-related code to get the x-axis shared animations
